backend/main_project/requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-json-widget==2.0.1
djangorestframework==3.15.2
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
mysqlclient==2.2.6
packaging==24.2
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
virtualenv==20.29.1
whitenoise==6.8.2
